+++ packages/hydrogen/src/components/index.ts
@@ -47,7 +47,6 @@ export {ProductMetafield} from './ProductMetafield';
 export {SelectedVariantPrice} from './SelectedVariantPrice';
 export {SelectedVariantBuyNowButton} from './SelectedVariantBuyNowButton';
 export {SelectedVariantShopPayButton} from './SelectedVariantShopPayButton';
-export {SelectedVariantImage} from './SelectedVariantImage';
 export {SelectedVariantUnitPrice} from './SelectedVariantUnitPrice';
 export type {
   BuyNowButtonProps,