+++ packages/hydrogen/CHANGELOG.md
@@ -135,6 +135,7 @@ function SomeComponent() {
 - feat [breaking change]: remove `assetHandler` parameter in the new `handleRequest`.
 - deps: update `react-helmet-async` to 1.2.3 and remove our custom types
 - dx [breaking change]: `<SelectedVariantAddToCartButton />` has been removed; the `<AddToCartButton />` will now use the selectedVariant by default.
+- dx [breaking change]: remove the `<SelectedVariantImage />` component in favour of using `<Image data={product.selectedVariant.image} />`
 
 ## 0.10.1 - 2022-01-26